Add error handling for IndexError on article instantiation.
Some papers on the arxiv do not have the appropriate fields filled for whatever reason, so they throw an IndexError during instantiation (the call to pop is on an empty array). Now we just skip those entries in read_arxiv so they don't gum up the process. This can probably be made more robust.